Finished implementation for the 'next day'
Implemented weather forecast speaking for the following day.
The weather forecast now tells a brief description of the weather,
and temperatures for the morning and the afternoon (sampled at 9 a.m. and
3 p.m.).
TODO: PROPERLY fix the issue of PROPERLY using the female_custom_voice.
For the moment being, the user can copy the file 'female_custom_voice'
to the directory 'espeak-data/voice' (obviously depending on the
installation and distro...) when possible. The scripts makes sure it
uses it in this case. Otherwise, the default voice is utilized.