Implemented init option
'init' option is now done. It allows the user to set up his environment,
typically when using the program for the first time.
That is, defining an API key to use in order to connect to
OpenWeatherMap and a defualt city ID to fetch the data for. The program
can now stores these pieces of info in dedicated files in the 'config'
directory. By the way, two 'feature' functions were added that can get the
user's location (either City and Country name, or Longitude + Latitude)
based on their IP address (which is not totally accurate I think but
correct to some extents). However, the main challenge is that the API
doesn't really behave well when trying to fetch data with a city name or
coordinates for the moment being. I hope this could be possible in the
near future, in order to enable the user to use the program without
being committed to a singleplace by defaut, especially if they frequently
change the location.